requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.1.0
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
whitenoise==6.11.0